database/requirements.txt
aniso8601==9.0.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
flask==3.0.3
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
flask-restful-swagger==0.20.2
flask-swagger-ui==4.11.1
importlib-metadata==8.2.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
pytz==2024.1
six==1.16.0
werkzeug==3.0.3
zipp==3.19.2
